+++ src/libtiled/properties.cpp
@@ -52,6 +52,11 @@ FilePath FilePath::fromString(const QString &string)
 
 void mergeProperties(Properties &target, const Properties &source)
 {
+    if (target.isEmpty()) {
+        target = source;
+        return;
+    }
+
 #if QT_VERSION < QT_VERSION_CHECK(5, 15, 0)
     // Based on QMap::unite, but using insert instead of insertMulti
     Properties::const_iterator it = source.constEnd();