+++ src/tiled/propertybrowser.cpp
@@ -414,6 +414,83 @@ void PropertyBrowser::wangSetChanged(WangSet *wangSet)
         updateProperties();
 }
 
+static bool isAutomappingRulesMap(const MapDocument *mapDocument)
+{
+    if (!mapDocument)
+        return false;
+
+    bool hasInputLayer = false;
+    bool hasOutputLayer = false;
+
+    for (const Layer *layer : mapDocument->map()->allLayers()) {
+        if (layer->name().startsWith(QLatin1String("input"), Qt::CaseInsensitive))
+            hasInputLayer |= layer->isTileLayer();
+        else if (layer->name().startsWith(QLatin1String("output"), Qt::CaseInsensitive))
+            hasOutputLayer = true;
+    }
+
+    return hasInputLayer && hasOutputLayer;
+}
+
+static void addAutomappingProperties(Properties &properties, const Object *object)
+{
+    auto addRuleOptions = [&] {
+        mergeProperties(properties, QVariantMap {
+            { QStringLiteral("Probability"), 0.0 },
+            { QStringLiteral("ModX"), 0 },
+            { QStringLiteral("ModY"), 0 },
+            { QStringLiteral("OffsetX"), 0 },
+            { QStringLiteral("OffsetY"), 0 },
+            { QStringLiteral("NoOverlappingOutput"), false },
+            { QStringLiteral("Disabled"), false },
+        });
+    };
+
+    switch (object->typeId()) {
+    case Object::LayerType: {
+        if (static_cast<const Layer*>(object)->name().startsWith(QLatin1String("input"), Qt::CaseInsensitive)) {
+            mergeProperties(properties, QVariantMap {
+                { QStringLiteral("AutoEmpty"), false },
+            });
+        }
+        break;
+    }
+    case Object::MapType:
+        mergeProperties(properties, QVariantMap {
+            { QStringLiteral("DeleteTiles"), false },
+            { QStringLiteral("MatchOutsideMap"), false },
+            { QStringLiteral("OverflowBorder"), false },
+            { QStringLiteral("WrapBorder"), false },
+            { QStringLiteral("AutomappingRadius"), 0 },
+            { QStringLiteral("NoOverlappingOutput"), false },
+            { QStringLiteral("MatchInOrder"), false },
+        });
+        addRuleOptions();
+        break;
+    case Object::MapObjectType: {
+        if (auto objectGroup = static_cast<const MapObject*>(object)->objectGroup())
+            if (objectGroup->name().compare(QLatin1String("rule_options"), Qt::CaseInsensitive) == 0)
+                addRuleOptions();
+        break;
+    }
+    case Object::TilesetType:
+    case Object::TileType:
+    case Object::WangSetType:
+    case Object::WangColorType:
+        break;
+    }
+}
+
+static bool checkAutomappingProperty(const Object *object,
+                                     const QString &name,
+                                     QVariant &value)
+{
+    Properties properties;
+    addAutomappingProperties(properties, object);
+    value = properties.value(name);
+    return value.isValid();
+}
+
 static bool anyObjectHasProperty(const QList<Object*> &objects, const QString &name)
 {
     for (Object *obj : objects) {
@@ -481,9 +558,10 @@ void PropertyBrowser::propertyRemoved(Object *object, const QString &name)
     if (!objectPropertiesRelevant(mDocument, object))
         return;
 
-    const QVariant resolvedValue = mObject->resolvedProperty(name);
+    QVariant resolvedValue = mObject->resolvedProperty(name);
 
     if (!resolvedValue.isValid() &&
+            !(isAutomappingRulesMap(mMapDocument) && checkAutomappingProperty(object, name, resolvedValue)) &&
             !anyObjectHasProperty(mDocument->currentObjects(), name)) {
         // It's not a predefined property and no selected object has this
         // property, so delete it.
@@ -1924,6 +2002,44 @@ void PropertyBrowser::updateProperties()
     }
 }
 
+Properties PropertyBrowser::combinedProperties() const
+{
+    Properties combinedProperties;
+
+    // Add properties from selected objects which mObject does not contain to mCombinedProperties.
+    const auto currentObjects = mDocument->currentObjects();
+    for (Object *obj : currentObjects) {
+        if (obj != mObject)
+            mergeProperties(combinedProperties, obj->properties());
+    }
+
+    if (isAutomappingRulesMap(mMapDocument))
+        addAutomappingProperties(combinedProperties, mObject);
+
+    const QString className = mObject->typeId() == Object::MapObjectType ? static_cast<MapObject*>(mObject)->effectiveClassName()
+                                                                         : mObject->className();
+
+    // Inherit properties from the class
+    if (auto type = Object::propertyTypes().findClassFor(className, *mObject))
+        mergeProperties(combinedProperties, type->members);
+
+    if (mObject->typeId() == Object::MapObjectType) {
+        auto mapObject = static_cast<MapObject*>(mObject);
+
+        // Inherit properties from the tile
+        if (const Tile *tile = mapObject->cell().tile())
+            mergeProperties(combinedProperties, tile->properties());
+
+        // Inherit properties from the template
+        if (const MapObject *templateObject = mapObject->templateObject())
+            mergeProperties(combinedProperties, templateObject->properties());
+    }
+
+    mergeProperties(combinedProperties, mObject->properties());
+
+    return combinedProperties;
+}
+
 void PropertyBrowser::updateCustomProperties()
 {
     if (!mObject)
@@ -1933,64 +2049,7 @@ void PropertyBrowser::updateCustomProperties()
 
     mCustomPropertiesHelper.clear();
 
-    Properties combinedProperties = mObject->properties();
-
-    // Add properties from selected objects which mObject does not contain to mCombinedProperties.
-    const auto currentObjects = mDocument->currentObjects();
-    for (Object *obj : currentObjects) {
-        if (obj == mObject)
-            continue;
-
-        QMapIterator<QString,QVariant> it(obj->properties());
-
-        while (it.hasNext()) {
-            it.next();
-            if (!combinedProperties.contains(it.key()))
-                combinedProperties.insert(it.key(), QString());
-        }
-    }
-
-    QString className = mObject->className();
-
-    if (mObject->typeId() == Object::MapObjectType) {
-        auto mapObject = static_cast<MapObject*>(mObject);
-        className = mapObject->effectiveClassName();
-
-        // Inherit properties from the template
-        if (const MapObject *templateObject = mapObject->templateObject()) {
-            QMapIterator<QString,QVariant> it(templateObject->properties());
-            while (it.hasNext()) {
-                it.next();
-                if (!combinedProperties.contains(it.key()))
-                    combinedProperties.insert(it.key(), it.value());
-            }
-        }
-
-        if (const Tile *tile = mapObject->cell().tile()) {
-            // Inherit properties from the tile
-            QMapIterator<QString,QVariant> it(tile->properties());
-            while (it.hasNext()) {
-                it.next();
-                if (!combinedProperties.contains(it.key()))
-                    combinedProperties.insert(it.key(), it.value());
-            }
-        }
-    }
-
-    if (!className.isEmpty()) {
-        // Inherit properties from the object type
-        if (auto type = Object::propertyTypes().findClassFor(className, *mObject)) {
-            QMapIterator<QString,QVariant> it(type->members);
-            while (it.hasNext()) {
-                it.next();
-                if (!combinedProperties.contains(it.key()))
-                    combinedProperties.insert(it.key(), it.value());
-            }
-        }
-    }
-
-    QMapIterator<QString,QVariant> it(combinedProperties);
-
+    QMapIterator<QString,QVariant> it(combinedProperties());
     while (it.hasNext()) {
         it.next();